[JBEAP-26556] Move some methods from JwtSecurityRealmTest into a new utility class
